requirements.txt
streamlit
pandas
requests
matplotlib
transformers
torch
openai
python-dotenv
google-generativeai
wordcloud
